Rename all project folders to match assembly names
This change renames all existing assembly folders to match their assembly
names.  In the new 'dotnet' project system, project references are done
using the package names of projects in the solution.  To keep the
project.json files consistent with what we expect in the NuGet packages,
it's necessary to ensure the project folders bear the assembly name so
that they can be referred to in the correct way.